fixbug callback android
