+++ android/src/main/java/com/livenessrn/LivenessView.kt
@@ -2,66 +2,64 @@ package com.livenessrn
 
 import android.content.Context
 import android.util.AttributeSet
-import android.util.Log
 import android.widget.FrameLayout
 import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.ReactContext
-import com.facebook.react.uimanager.ThemedReactContext
-import com.facebook.react.uimanager.UIManagerHelper
-import com.facebook.react.uimanager.events.Event
+import com.facebook.react.bridge.WritableMap
 import com.facebook.react.uimanager.events.RCTEventEmitter
 import com.liveness.sdk.core.LiveNessSDK
 import com.liveness.sdk.core.model.LivenessModel
 import com.liveness.sdk.core.utils.CallbackLivenessListener
-import com.livenessrn.events.RNLivenessEvent
 
 
 class LivenessView @JvmOverloads constructor(
   context: Context,
   attrs: AttributeSet? = null
 ) : FrameLayout(context, attrs) {
-  private val themedReactContext: ThemedReactContext = context as ThemedReactContext
-  private val TAG = "LivenessView"
-  private val EVENT_NAME = "onEvent"
-
   private val callBack = object : CallbackLivenessListener {
-    override fun onCallbackLiveness(data: LivenessModel?) {
-      Log.d("callBack", "data==== $data")
-      val map = Arguments.createMap()
+    override fun onCallbackLiveness(livenessModel: LivenessModel?) {
+      if (livenessModel != null && livenessModel.status != null && livenessModel.status == 200) {
+        val map = Arguments.createMap()
+        map.putInt("status", livenessModel.status ?: -1)
+        map.putString("message", livenessModel.message ?: "")
+        map.putString("request_id", livenessModel.requestId ?: "")
+        map.putString("code", livenessModel.code ?: "")
+        map.putBoolean("success", livenessModel.success ?: false)
+        map.putString("pathVideo", livenessModel.pathVideo ?: "")
+        map.putString("faceImage", livenessModel.faceImage ?: "")
+        map.putString("livenessImage", livenessModel.livenessImage ?: "")
+        map.putString("transactionID", livenessModel.transactionID ?: "")
 
-//      LivenessRnPackage.sendEvent(context as ReactApplicationContext,"onEvent", map)
-      callBackEvent(data)
+        if (livenessModel.data != null) {
+          val mapData = Arguments.createMap()
+          map.putString("faceMatchingScore", livenessModel.data?.faceMatchingScore ?: "")
+          map.putString("livenessType", livenessModel.data?.livenessType ?: "")
+          map.putDouble("livenesScore", (livenessModel.data?.livenesScore ?: 0).toDouble())
+          map.putMap("data", mapData)
+        }
+        callNativeEvent(map)
+      } else {
+        val map = Arguments.createMap()
+        map.putInt("status", livenessModel?.status ?: -1)
+        map.putString("message", livenessModel?.message ?: "")
+        map.putString("code", livenessModel?.code ?: "")
+        callNativeEvent(map)
+      }
     }
   }
 
   init {
-    Log.d("callBack", "data==== 1122122112")
     LiveNessSDK.setCallbackListener(callBack)
-    emitEvent("A===========================")
   }
 
-  private fun callBackEvent(data: LivenessModel?) {
+  fun callNativeEvent(map: WritableMap) {
     val reactContext = context as ReactContext
-    reactContext.getJSModule(RCTEventEmitter::class.java)
-      .receiveEvent(id, LivenessViewManager.EVENT_NAME, Arguments.createMap().apply {
-        putString("text", data?.status?.toString())
-      })
+    val event = Arguments.createMap()
+    event.putMap("data", map)
+    reactContext.getJSModule(RCTEventEmitter::class.java).receiveEvent(
+      id,
+      "nativeClick",  //name has to be same as getExportedCustomDirectEventTypeConstants in MyCustomReactViewManager
+      event
+    )
   }
-
-  private fun emitEvent(data: String?) {
-    val reactContext = context as ReactContext
-    val id = id
-    val dispatcher = UIManagerHelper.getEventDispatcherForReactTag(reactContext, id)
-    val surfaceId = UIManagerHelper.getSurfaceId(reactContext)
-    if (dispatcher != null) {
-      dispatcher.dispatchEvent(RNLivenessEvent(surfaceId = surfaceId, id, 112))
-    }
-  }
-
-//    val reactContext = context as ReactContext
-//    reactContext.getJSModule(RCTEventEmitter::class.java)
-//      .receiveEvent(id, LivenessViewManager.EVENT_NAME, Arguments.createMap().apply {
-//        putString("text", data)
-//      })
-//  }
 }
