+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "liveness-rn",
-  "version": "1.0.31",
+  "version": "1.0.32",
   "description": "My Liveness Library",
   "main": "lib/commonjs/index",
   "module": "lib/module/index",
